mini_app/backend/requirements.txt
fastapi[standard]
psycopg[binary,pool]
python-dotenv
alembic
SQLAlchemy